Add a disconnection timeout to emit expired url signal
This timeout is used in situations where client disconnects when a media
transmission is taking place and he or she does not connect any more and the
cookie timeout lasts longer than needed to notify the java application. If
timeout is configured to be short enough, java application can realise that
client has disconnected before cookie expires and thus cancel the media
transmission.

Change-Id: Iae66b9ac22ff9b20c35728c818bff2f10af55c0b