audiomixer: Create source pad as soon as it is requested a sink pad
Do not wait until media flows in sink pad to make internal reconnections
in audiomixer. This patch makes internal reconnections as soon as
a sink pad is requested and produces a source pad that mixes all other
audio entries in it excepting the one is connected. On the other hand,
when media starts flowing in the sink pad, it is reconnected to all other
live adders causing other existing outputs to have the new audio stream
mixed in it.

Change-Id: I3ab6f9e7d1d2e28925ce9efe452b8249613865c2