Distributed 1 to N video call
This example uses three pipelines to deplay the video call application.
The main pipeline is where the WebRTCEnpoint master will be created. The
two other pipelines will be used to create the WebRTC peer endpoints. Such
peers will be created in one or another pipeline as they connect to the
application to distribute the overhead among other mediaservers attending
requests to such pipelines.

Change-Id: I9196ac09a8cfb082ec00bdd88f00278421781ddd