SCTP: Disable Nagle algorithm to speed up data transmissions
This means that data is generally sent as soon as possible and no
unnecessary delays are introduced, at the cost of more packets in
the network.  When this option is set to zero, data transmissions
may be temporarily delayed waiting for additional data to be
transmitted in the same packet.

Change-Id: I08c9956cfd32441c61228d9b62e48915fdec54bd