Http: Fix wrong memory access destroying singleton httpServer class
HttpEndPoint keeps a server reference to their server. Now, server
keeps alive as long as there are alive HTTP end points. No matter
whatever the destruction order of singleton classes are. When last
reference to the singleton http server is destroyed, the server
object will be stopped and destroyed as well.

Change-Id: I8d72b750f8af9ec36595fde9638fc60eb3a06c6a