kmswebrtcbaseconnection: add get_certificate_pem
erdtlssrtpdec from Ericson already generates a certificate so if we
want to use a custom one we must provide it as a string containing
a X509 certificate and RSA private key in PEM format. So far we use
the one generated by erdtlssrtpdec and provide the method to get it.

Change-Id: I7551d4369fbd5075f74ca601f363fe8ef82bf6c9