kmsplayerendpoint: Perform a seek if pause state change return success
If state change to pause returns success the paused is not performed
due to the configuration of the elements in the internal pipeline. For
that reason, if we get a success we perform a seek to the current position
and then we perfom a state change that returns async. This hack only
executes the first time that we pause the pipeline.

Change-Id: I0a39d002c1a8b5995cce65a76793d069b3c923c0