compile_project.py: re-enable timestamps in Release versions
We need to think about a better workflow that integrates the various
needs from development, CI and deployment points of view.
Currently we have a linear pipeline of jobs in Jenkins, which requires
building the same versions multiple times, and this will fail because
the package repos don't accept an already existing file name.