Fix test: test_not_enough_ports, depending on libnice version
libnice 0.1.13 should fail in the test case "test_not_enough_ports",
because the second offerer cannot get two UDP ports for its two
components.

However, libnice 0.1.14 was improved in this regard, and won't fail
because even if it doesn't find UDP candidates, it should be able to
find TCP-ACTIVE type ones. In this case, the test should check that not
all of them are UDP.