Python print(): pass list of objects instead of string concatenations
print("My variable: " + var) passes a single string object to print(),
but the string concatenation will fail if 'var' is not a string.

print("My variable:", var) passes two objects to print(), which get
passed through the str() function, so the call won't fail even if 'var'
is not of string type.