[REVIEW] Go back to using 'curl' for uploads
CURL sends an initial empty request, with Header "Expect: 100-Continue"
This makes Apache reply with a status 100, and it is then that CURL
starts transferring the actual binary payload.

For some reason, with this method Apache doesn't honor the upload size
limits (parameter 'LimitRequestBody'), and it allows the upload without
problem.
However, any other HTTP client, such as Python Requests, doesn't perform
this trick, and so the server rejects the upload with the status "413
Request Entity Too Large".

The sysadmin refuses to investigate this misconfiguration on the server,
so the only option is to go back and use the 'curl' command.

Related:
- https://bz.apache.org/bugzilla/show_bug.cgi?id=39243#c40
-
https://github.com/requests/requests/issues/3299#issuecomment-224973180