Player/HttpEndpoint: With 'useEncodedMedia', don't accept RTP
By setting the "caps" property to the uridecodebin element, with all
formats except 'application/x-rtp', what we achieve is that all incoming
formats will be passed directly to the media pipeline (as is expected of
the 'useEncodedMedia' mode), but incoming RTP streams will still be
depayloaded.
Passing RTP packets directly to the pipeline, without depayloading, is
not supported.