Tests: Increment Check timeout to 60 seconds
The libsoup library has already a timeout of 60 seconds for its
operations, so in the worst case it could finish a request right before
the Check framework decided to abort the test.

With 120, worst-case HTTP requests can take a long time and the test
will still have time to run.