Debian: Remove 'override_dh_auto_test' from debian/rules
3 reasons:

1. Forcing to run 10 tests in parallel could cause CPU exhaustion
problems. Currently, CI runs with 4 CPU cores, so the ideal job count is
4.

2. 'dh_auto_test' already runs with as many jobs as possible, because we
build packages with `dpkg-buildpackage -jN`, where N = 4. This appends
internally the appropriate environment variable for Debhelper:

    DEB_BUILD_OPTIONS="parallel=4 $DEB_BUILD_OPTIONS"

In the end, all this translates implicitly into this Makefile rule,
which is practically the same that we have now in our override:

    dh_auto_test: make -j4 test ARGS+=-j4

3. 'override_dh_auto_test' does NOT honor the config option "nocheck".
This means that tests couldn't be disabled by exporting the environment
variable DEB_BUILD_OPTIONS="nocheck" before building the packages.

This is a bug in Debian, see #568897
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=568897