kurento_maven_deploy: Move 'package' goal back with 'deploy'
The maven-deploy-plugin needs to receive the build metadata from the
previous goals such as 'package', or else it won't be able to know which
files are expected to be deployed.

Having a separate 'package' command will break this dependency, so
'package' and 'deploy' must be brought back to the same command
invocation.