Rename dev-media-server and release-media-server
dev-media-server -> media-server-buildtools-nightly
release-media-server -> media-server-buildtools-release

This way both dirs (and CI jobs) have more consistent names and
are easier to locate.

Also, in the future these are meant to be documented so users
are able to build packages from their own module directories. Having
clear names is a part of the documentation effort.