ci_container_job_setup: Force Docker ENTRYPOINT to /bin/bash
The Docker ENTRYPOINT and CMD was inherited from base Ubuntu images, and
it was '/bin/bash' anyway. But now we're offering images
(kurento/buildtools) with custom ENTRYPOINT values so they function as
if they were executable programs, to be used by end users.

This is cool, but our CI scripts expect that the default command is
'/bin/bash', not some random script. So, we force it here.