New 'kurento-buildpackage' and 'kurento-ci-buildtools'
The 'kurento-buildtool' image has been simplified to not install NPM and
other tools which are not required to build Debian packages, and renamed
to 'kurento-buildpackage' because that's the main purpose of this image.
'kurento-buildpackage' is meant to be used by Kurento devs AND USERS to
build packages.

All those tools are then moved to a new image, 'kurento-ci-buildtools'
which is meant to be used as base image for all CI jobs related to
Kurento. 'kurento-ci-buildtools' are meant to be used ONLY BY KURENTO
CI, to run all C/C++, Java and Javascript jobs.