New script: kurento_git_checkout_name.sh
Check out a given branch or tag name, if it exists.

This script will try to check out the provided name in the current Git
repository, and will revert back to the name 'master' if the desired one
does not exist.

Arguments
---------

<Name>

Branch or tag name that should be checked out.