New script: kurento_ci_aptly_repo_create.sh
CI - Create or update Aptly repos to store Debian packages.

This script will handle all Aptly operations to create or update already
existing repositories.

The script is meant to run in the remote server that hosts Aptly repos.
It won't work if you run this script locally or in the Jenkins machine;
instead, you should copy it to the target server and run remotelly via
SSH.