test/common: Add fatal errors when Docker operations fail
Having so many test jobs makes it impractical to only have a
console.error() debug message in the logs. We need the job to actually
fail if there is any problem with these operations, or else nobody will
notice them and the docker containers will be left around, making future
jobs to also fail.