generate_js_module: Patch random sorting of 'remoteClasses' array
FIXME - When generating for the kms-filters module, the JSON keys in
'js/src/filters.kmd.json' change ordering each time this job runs. So
the git commit history is polluted with meaningless variations in key
order.
I haven't found the cause for this issue, and it is really a time sink
without much payoff, so the easy way is to force a sorting order right
here