FIXME: Disable test 'player_with_filter' temporarily
Test disabled during the migration to Ubuntu 18.04 (Bionic), because the
CURL plugin is not built for gstreamer-plugins-bad, and that means the
PlayerEndpoint doesn't have support for "http" URIs, which is needed to
load the sample file used in this test.
The test will fail and that prevents building packages for this module.