mavenize_js_project: Fix ZIP assembly with JS resouces
Files auto-generated by kurento_mavenize_js_project.sh include an
assembly descriptor, to generate a ZIP with all JS files. However, this
wasn't being called, so no ZIP was being generated.

This commit enabled the generation of ZIP file. It was either removing
the code or actually making use of it. For now, let's use it, and if
later we confirm there is no need for the ZIP files, we can delete all
this.