WebRtcEndpoint: Change "externalAddresses" to "networkInterfaces"
(Hopefully) last iteration of the work started in
* Kurento/kms-elements@5619b48a1d4ed69d4699feb07ee1fde1c921b9fe
* Kurento/kms-elements@32adf2ed99976d9e39a4713cd3325b775a5d0b49

In this commit we change our minds; "externalAddresses" was a misleading
name because the parameter didn't really allow to set an external IP.
Only local IPs were accepted (by libnice) in order to open listen sockets,
so discovery of the actual external IP address was needed.

GStreamer-plugins-bad is working on adding a similar parameter, in
webrtcbin: add ice-nics property
https://gitlab.freedesktop.org/gstreamer/gst-plugins-bad/merge_requests/164

This is a good idea, because in virtual environments it is easier to know
the names of the network interfaces, than the IP addresses that they
will end up being assigned.

This commit implements that change, so now it should be possible to
indicate which network interfaces should be used by Kurento (libnice
under the hood)