WebRtcEndpoint: Add "externalAddress" to config & client API
Allows to specify an external IP address, which will be used to mangle
all gathered local ICE candidates before sending to the remote peer.

This way, remote peers are able to know about the exteral or public IP
address of Kurento, without the need to install and configure a STUN
server for Kurento.