kmsicecandidate: Don't set VALID when mDNS cannot resolve
When mDNS resolution is enabled in the system, reception of Chrome's
mDNS ICE candidates works fine.

However, in environments where mDNS is not available:

* Docker images
* Ubuntu systems where the package libnss-mdns is not installed
* Systems where /etc/nsswitch.conf doesn't contain mdns4_minimal in the
list of host resolution engines

the resolution of Chrome's mDNS candidate will fail, as expected, but
the candidate itself will be considered valid. Considering it valid
means that it is processed, stored in the queue of candidates to be
added to libnice, and eventually added.

libnice rejects adding an mDNS candidate because it cannot parse it, so
it returns an error that ends up unwrapping as a Kurento API Error
message, and this will make most applications abort because errors are
not expected, even though invalid mDNS candidates is a perfectly valid
situation.

The solution is to simply reject mDNS candidates in the case their
address cannot be resolved.

Fixes Kurento/bugtracker#432