kmsrtpendpoint: Correct parsing of "a=rtcp" attributes
The code was first removing the attribute, then adding a new one. This
means that the loop logic would step over the next attribute that might
come next, because the index variable gets incremented, while the array
of attributes has been modified in place.

This changes the attribute struct in place, so the loop logic is not
altered by the modification.