WebRtcPeer: Manually add transceivers only for "recvonly"
Otherwise, duplicated medias will be offered in the SDP Offer.

Report: https://groups.google.com/d/topic/kurento/LJq3YbNzyVo/discussion

The problem is caused because pc.addTransceiver() is called
unconditionally on all calls to generateOffer, regardless of the mode of
operation (sendrecv, sendonly, recvonly), here.

This is incorrect, because in cases where we'll be sending (sendrecv,
sendonly), the WebRtcPeer constructor has already created tracks as part
of the GetUserMedia action, here, and then here.

So, creating new recvonly transceivers in generateOffer should be only
done in the case of mode == recvonly, which is the only case that hasn't
yet been covered by the code at that point.

Fixes Kurento/bugtracker#431