kurento-media-server: Don't fail on accessing /etc/hosts
If the container is run with an user which is not root, it won't have
permission to write /etc/hosts and the command will fail.

Add "|| true" to prevent the exit error. This will leave the IPv6
loopback without disabling, of which we should study what potential
issues this may cause in modern Kurento versions (>= 6.13), if any.