consigure.sh: Set version as "master" for non-release builds
When the documentation build is for a release, we want the tutorials to
instruct the user to check out the correct version that corresponds to
the tutorial text; so, it the docs are for version 1.2.3, we want the
text to read:

git checkout 1.2.3

However for all nightly builds of the docs, users should be using the
latest version of the tutorials, because there may be some issues fixed
that haven't been released yet. So, a user reading the "latest" version
in ReadTheDocs should see this:

git checkout master

Ideally she should not see "git checkout" at all, but that means we'd
have to use conditionals, which Sphinx doesn't seem to support. The
extension "ifconfig" is not good to be used inside source code blocks.