webrtcendpoint: unify log messages
Use the names of the Client API events to group different debug messages
with the same prefix. This way, it is easy for a user to identify which
log messages correspond to the "action" that is related to the given
event.

There are 3 levels at which the ICE messages can occur, from internal to
external:
* libnice agent control (kmsiceniceagent)
* WebRTC session handling (kmswebrtcsession)
* WebRtcEndpoint implementation (kmswebrtcendpoint)

The most internal ones are set to DEBUG level, while the endpoint
implementation is the one providing information to the user by default,
and is set to work on the INFO level.

With this, the default log levels for a Kurento installation contain all
details about the ICE candidate gathering, and all messages come from
the same source (kmswebrtcendpoint).

The only exception is from Peer-Reflexive ICE candidates, which are
detected and processed only at the most internal layer
(kmsiceniceagent), so these are logged at the INFO level.