all: explain that "wss://" requires installing a Root CA
Browsers allow doing some specific tests with self-signed certificates,
such as accessing HTTPS pages (they will show a warning about insecure
certificates, but will otherwise allow accessing the page).

However, browsers will not be happy to connect a Secure WebSocket that
is served with a self-signed certificate. To do so, you need to either
use a proper certificate (such as generated by Let's Encrypt), or
install the self-signed Root CA into the browser's trusted store.

To do the later, the took mkcert is very useful and easy to use:
https://github.com/FiloSottile/mkcert