webrtcendpoint: fix leaked sockets from libnice breaking change
nice_agent_remove_stream(), called from
kms_ice_nice_agent_remove_stream(), was silently converted from
synchronous to asynchronous in libnice 0.1.16. I know that Semantic
Versioning tells us that 0.X.Y should be considered as unstable API, but
come on, that change wasn't *nice* of libnice ;-)

To solve the issue, we need to run a last iteration of its GMainLoop
context in order to allow it release all resources and object
references. Otherwise, libnice takes a reference to the NiceAgent object
in nice_agent_remove_stream(), and it never gets to release it, because
the unref is done in the callback function on_stream_refreshes_pruned().

Thanks to @omgold for finding the root cause of the issue and proposing
a solution!
https://github.com/Kurento/bugtracker/issues/460#issuecomment-619529496

A bug report has been filled in libnice as a result of this commit:
https://gitlab.freedesktop.org/libnice/libnice/issues/110

Fixes https://github.com/Kurento/bugtracker/issues/253
Fixes https://github.com/Kurento/bugtracker/issues/460