all: change to non-secure WebSocket by default
Users were complaining that JS tutorials didn't work. It turns out that,
since KMS defaults to non-secure WebSocket, using Secure WebSocket by
default in tutorial code was not too convenient.

In order to use Secure WebSocket, users need to configure KMS
adequately, following an extra set of instructions. We want these
tutorials to be as easy and smooth to use as possible, so removing
intermediate steps is desirable to achieve a "works out of the box"
experience.