ci_job_generate_module: don't build for experimental branches
No need to install all .deb packages, which takes a long time, only to
then skip building the JS or Java module because we're working on an
experimental branch (because we don't build API modules from
experimental branches).

So, skip all the work and exit right away as soon as possible.