Use "ws://{KMS_HOST}:8888" in all examples
Using the default port "8888" is a good way to make it clear that
insecure WebSocket is being in use (ws://). When Secure WebSocket is
used, the URI scheme is changed to "wss://" and the port must be
accordingly changed to 8433 (by default).

All other sources and examples were already using 8888 or 8433, so it is
consistent if we use these also for "mvn" example commands.

Also, break long "mvn" command lines so they aren't too long.