ci-buildtools: install a well tested NPM version
Otherwise the default NPM version from Node 8.x will get installed,
which has a bug that breaks cross-mountpoint file accesses during 'npm
install'

It is probably safe to install newer versions of NPM in the future, but
for now we'll use a specific, pinned version that we know works fine.

The NPM versio nwill then be updated manually by a conscious decision,
not automatically whenever there is something new pulling it.