tests/webrtcendpoint: use a HashMap to store gathered addresses
The previous method of relying on a limited number of ports was working
only for IPv4. As soon as IPv6 is used, this was confusing and prone to
errors. Also it failed because with IPv6 it was actually finding an UDP
candidate for the RTCP component (number 2) of the second
WebRtcEndpoint.

This method is more robust: just store all seen candidates, and check
that they never get to repeat. It does the same as before, but it works
better for both IPv4 and IPv6.