kurento-buildpackage: use "--no-remove" to avoid silent removals
With --no-remove we avoid a situation where a bad package configuration
might result in a successful build, because the conflicting packages
would be uninstalled prior to the build step.

The issue would then manifest later, when trying to install the
conflicting packages. Better to fail fast and trigger an error as soon
as possible!