templates: add comment about constructor params
For custom GStreamer style modules, the FilterImpl constructor expects a
MediaObjectImpl object, of which mediaPipeline is one if the appropriate
dynamic cast is applied.

Instead of leaving the developer wondering about that, give an explicit
hint about this in the TO-DO comment.

Also reword all other comments, they are not FIXMEs (in the sense of
something is broken, which is not), they are TO-DOs