kurento-buildpackage: change base buildpack-deps -> ubuntu
We're trying to get rid of the curl package in our images, because it
depends on libcurl which is currently causing incompatibility issues for
the Ubuntu Bionic builds. Same with CMake.

So let's use a plain ubuntu base image without basic tools that will get
installed anyway by those projects which actually depend on them.