kurento-buildpackage: HACK for Ubuntu 18.04 Bionic: install libcurl
Until migration to Bionic is completed, we have to keep the code working
in both 16.04 and 18.04 at the same time. This imples installing both
libcurl3 (OpenSSL 1.0) and libcurl4 (OpenSSL 1.1) at the same time...
which is impossible because they conflict.

Gicen that libcurl4 is only needed as a dependency of CMake, but then we
NEVER use any download feature of CMake, we add this HACK in the script:
edit both libcurl* packages, and remove the "Conflicts:" statement, so
it is possible to install both at the same time without conflics, thus
satisfying the dependencies of cmake and other development packages.

This can be reverted once support for Ubuntu 16.04 Xenial is dropped,
and all our code is updated to work with OpenSSL 1.1.