Revert "parent-pom: Enable openvidu repository for both snapshots and releases"
This reverts commit 966b5b166b14a9166957ca1449d162071405b8cf.

Under no circumstance we should leave the SNAPSHOT repository enabled
for Release builds. This caused that some SNAPSHOT dependencies slipped
into the Maven build tree.

If versions-maven-plugin is not updating its own version to SNAPSHOT
automatically after a release, we'll need to change the version update
mechanism, or just write down the new version by hand. But enabling the
snapshot repo for releases is NOT the solution.