kurento-buildpackage: add '--apt-add-repo', '--apt-proxy'
These options allow to fine tune the calls to apt-get.

'--apt-add-repo' is a BREAKING CHANGE: from now on, '--install-kurento'
WILL NOT add the requested apt repositories if they are not already
present in the system. This is being changed because editing the user's
system settings without explicit consent is not a nice thing to do.

By calling with '--apt-add-repo', we consider the user is conscious
about the changes that this tol will make to their system.