fixup! kurento-buildpackage: change base buildpack-deps -> ubuntu
This partially reverts 26094e74bbe77fc0980f4c53856614d1d756cae3

It was a mistake to remove the build-essentials package: it is needed by
jobs that will auto-generate Java and JavaScript source code from
Kurento modules that contain KMD files.

build-essential is not defined in all debian/control files as a build
dependency, because the Debian packaging tools assume by default that
this package is necessary, and also it is a linter error to depend
explicitly on that package from the debian/control file.