parent-pom: remove dep: kurento-tree
This project has been neither maintained nor supported since 2016.
At most, it might have received minor corrections over time, to fix
compilation errors.

It makes no sense to have this as a mandatory dependency of parent-pom,
so in this commit we set this project to a fixed version and leave it
there in case work on it resumes in the future.