merge_doc: use rsync instead of cp
To copy all files including possibly new hidden ones, use rsync
instead of cp. This gives more flexibility to set exclude rules. The
'--delete' already takes care of deleting old files in the target dir
which have been removed from the source dir.

--exclude-from=.gitignore allows easily excluding the

This was prompted by the addition of .readthedocs.ymlm, which is the new
recomended way to specify build configs for Read the Docs.