drop "json-object" blocks: deprecated in Pygments 2.8
Given comments in https://github.com/pygments/pygments/issues/1600, the
json-object lexer is geting deprecated with the upcoming Pygments 2.8,
and will fall back to "json".

At the same time, Pygments 2.8 comes with a new JSON parser that allows
for partial blocks missing surrounding curly braces, which is exactly
the use case we have in this documentation. Nice update!