set-versions: don't enforce that old versions start with digit
This was a "just in case" check, verifying that the old version number
started in all cases with a digit. But, it's better to be more lax and
allow basically anything, given that we're going to replace it
completely sol it doesn't matter if the old value was a correct one.