recorder: add config to choose gaps fix method (#31)
## What is the current behavior you want to change?

Allow choosing between the original broken behavior or the recorder (keep gaps) and the new one introduced in Kurento 6.15.0 (remove gaps in recordings): https://doc-kurento.readthedocs.io/en/6.15.0/project/relnotes/v6_15_0.html#fixed-recorder-synchronization-on-streaming-gaps

## What is the new behavior provided by this change?
* New CONFIG FILE: **RecorderEndpoint.conf.ini**
* New CONFIG OPTION: **gapsFix = (NONE | GENPTS)**
* DEFAULT setting: `gapsFix=NONE`

## How has this been tested?
Debug log to show that the feature is enabled. Not really tested yet (will publish experimental packages for it)

## Types of changes

Kurento 6.15.0 changed the recording behavior because a recording with gaps in it should be considered a bug, thus an invalid file. So the breaking change was allowed within minor versions because it was considered as a bug fix.

But XKCD 1172 came to bite us because we need that the total duration of the recordings matches with the total duration of the sessions that were recorded, which does not happen if you remove timestamp gaps from recordings. So we need to roll back the change.

Doc changes already included in the PR.