kurento-media-server: rename custom apt-get() to apt-get-install()
Better not to alias the default command, and make it clear that this is
a custom command on top of apt-get.

The previous name could be very confusing for people reading the
Dockerfile, as they would see "apt-get install" and think that it was a
plain and simple call to the apt-get binary, when in fact it was calling
our custom bash function.