kurento-media-server: FIXME repeat apt-get until it succeeds
Our OpenStack network is causing spurious disconnections while
downloading apt packages, which causes a failure due to a bug in
apt-get (which wasn't fixed until Ubuntu 20.04).

Also, this problem happens when using the OpenVidu package
proxy, but in other jobs (kurento-media-server-asan) the same problem
occurs when _not_ using the proxy.

So this quick and dirty solution simply removes usage of the proxy and
loops again and again until the command succeeds. This should end up
happening eventually, because the apt cache is populated and reused
between sequential invocations of the apt-get command.