mark "development" scripts for deletion
The "Development" folder was intended as a place where to put small
utils that could be useful FOR DEVELOPERS, to run in their systems for
miscellaneous operations like installing or uninstalling Kurento
packages.

However in practice these are neither used, nor maintained, so they have
all fallen into oblivion. Better to just get rid of them.