AWS CloudFormation: Use the Init stage to run setup script
Use the UserData section for just the minimum possible amount of setup.

The UserData section uses the "!Sub" macro, which replaces all AWS-
specific variables (such as e.g. ${AWS::StackId}) with their actual
value. The problem is that this syntax collides with the standard shell
syntax for env var expansion!

This had a solution, but an ugly one: the shell expansions had to be
"escaped" and look like this: ${!VarName}, a non-standard syntax that is
particular to CF.

A much more elegant and maintainable solution is to have all initial
setup comands in their own script (cfn-setup.sh), and call this script
from the initialization stage of the EC2 instance, with `cfn-init`.