AWS CloudFormation: remove unused Init settings
These lines were not being used. I know because of 2 reasons:

* The syntax is bad. All these settings should have been inside a
parent "config" key, as the docs clearly explain.

* I verified it by looking at the log files (/var/log/cloud-init.log,
/var/log/cloud-init-output.log) and no actions were performed
precisely because the key "config" was not present.

In any case, even if this was correct we can delete it because:

* We are not using `aws-cli` for anything.
* We are not using `cfn-hup` for anything.
* There is nothing at this path: /opt/aws/bin/cfn-init
so even if this runs, nothing will be actually done anyways.

Feel free to blame me if this ends up breaking something.