webrtc, rtp: replace G_MAXUINT16 with 65535
Using G_MAXUINT16 was semantically confusing. It conveys something like
"this value is limited by the range of 16bit integers", when that's just
a coincidence.

We want to say "max port as defined by Linux", so if tomorrow they
change to 65539, then that's the new value to use. The fact that Linux
happens to use MAXUINT16 as their max port number, is just an
implementation detail on their part.