writing_doc: rewrite more concise, use venv
Sometimes I get carried away and write walls of text for something that
could have been a couple lines...

This should be easier to read, even skip quickly, and easier to follow.

Also it adds usage of Python's virtual environment, which is a standard
practice in Python's world instead of installing packages globally.