maven_deploy: don't force "default" profile
When deploying snapshot versions, it makes no sense to activate both
"deploy" and "default" profiles of "kurento-java". Both of these will
try to build and deploy multiple times the same submodules, causing a
conflict in the GitHub registry.

"default", also, includes running "kurento-assembly" and
"kurento-basicroom", which we don't want deployed.