parent-pom: fix duplicate executions of maven-source-plugin
Follow advice from https://stackoverflow.com/a/10794985

Best practice says that to get consistency you need to configure this in
the root POM of your project in build > pluginManagement and NOT in your
child poms. In the child pom you just specify in build > plugins that
you want to use maven-source-plugin but you provide no executions.