parent-pom: remove snapshot repositories
Docs have been updated to show how to enable the new GitHub packages
repository. Including a direct reference to the snapshots repo in the
pom file has proven to be problematic, because even for Release builds,
Maven does access the snapshots repo and downloads all metadata for all
available nightly versions!!!

So, really, snapshot repos should be defined inside a profile in the
settings.xml file, and be enabled on purpose with `-P`. That's what
we'll be doing now, so the repositories section here should be deleted.