pom.xml template: set as child of "kurento-parent-pom"
The Kurento Java API artifacts were independent, but this meant they
lacked access to all the common configuration that is carefully
maintained in the kurento-parent-pom.

Specifically, the <distributionManagement> section of the parent pom,
which contains repository definitions that are used by `mvn deploy` to
upload release or snapshot artifacts to the appropriate repos.