parent-pom: use correct property for parent-pom version
${project.version} is not what we really want to use here, because in
real projects this gets resolved with the children's version, not the
parent pom's one.

So if my project, version 1.2.3, is a child of parent-pom, all mentions
of ${project.version} in the parent-pom become "1.2.3", which is
obviously wrong.

The solution is to use ${project.parent.version}; then, whenever a
project declares this as its parent pom, the placeholder will be
expanded to the actual version of the parent pom, regardless of the
children's one.