prepare for siprtp_module (#32)
Current RtpEndpoint (GStreamer plugin) is not well suited for SIP RTP media integration, as it lacks several features. This Lacking features are implemented in an external module SipRtpEndpoint that regarding its interface it derives from RtpEndpoint (available in https://github.com/naevatec/kms-siprtpendpoint). However, current RtpEndpoint plugin declares several data private that is needed on derived SipRtpEndpoint in that module. In fact current RtpEndpoint plugin can be considered "final" in the sense that it is not intended for deriving any class. The alternative would have been to reimplement many features of RtpEndpoint in SipRtpEndpoint, which is for sure a bad idea in terms of maintenance.
So, this PR is about re-engineering a little part of RtpEndpoint plugin so that private data gets public and so can be reused by derived classes/plugins.