cmake: Disable modern Boost imported target from newer CMake
Boost >= 1.70 started to provide a CMake package configuration file for
use with `find_package()`. This defaults to imported targets, which is
actually the modern CMake way to use packages.

However, Kurento was written for earlier versions, and expects full
paths to boost libraries instead of just the target aliases (such as
"Boost::regex").