webrtc tests: Delete 'test_not_enough_ports'
This test was actually testing an implementation detail of libnice; as
such, any change in the libnice internals could make the test fail.

And indeed that's what happened; a change between libnice 0.1.17 and
0.1.18 modified the way how the library manages a "not enough ports"
scenario; and actually, for the next version, the author has changed his
mind again and it will change to something different when 0.1.19 is
released.

So better just to ignore this behavior, as it is a detail of the
library, it should be tested by the unit tests of the library.