pom.xml template: Add <distributionManagement> for uploading packages
Add standalone <distributionManagement> section with the Release and
Snapshot repositories for Kurento. These will be used to upload Release
and Snapshot artifacts, respectively.

It was not a good idea to depend on kurento-parent-pom to obtain the
<distributionManagement> section from there, because of a
chicken-and-egg problem; the next best (and quickest) solution seems to
be to just add this section in generated pom.xml files.