get_version: Use help:evaluate instead of exec:exec
exec:exec tries to download and install all of the project's
dependencies, even if it is just to print the project's version. This
fails when the dependencies aren't available (e.g. if they are
snapshots, because the snapshot profile is not being enabled here).

help:evaluate doesn't need to install dependencies. So even if there is
some dependency missing, it will work and print the project version.