pom.xml: Remove maven-deploy-plugin
CI scripts call the deploy plugin with a fully qualified name, to
control the exact version of the plugin that gets used. As a
consequence, there is no need to specify the plugin in pom.xml, as the
"deploy" phase is never called as part of the Maven Build Lifecycle.