templates: Don't replace all occurrences of "lib", use regex "^lib"
Files were generated with wrong names if the module name contained "lib"
anywhere in its name. For example, for a plugin named "MyCvModuleCalib":

$ mkdir build/ ; cd build/
$ cmake ..
[...]
-- Generated: kmsmycvmoduleca.pc.in
-- Generated: FindKMSMYCVMODULECA.cmake.in
```

This is because any appearance of the word `lib` was being removed from
file names. Probably an attempt to convert things like `libsomething`
into `something`, but with the unintended effect of converting `mycalib`
into `myca`.

Solution is to use FreeMarker's ability to match by regex, and only
replace "lib" from the beginning of the name. I'm not sure that's
correct either, but it has been like that for years so I'm hesitant to
remove the replacement completely.

Fixes https://github.com/Kurento/bugtracker/issues/289