Revert "kmd: Remove dependency on kurento-client for generated pom.xml"
This reverts commit 4391fad6b5e724d144440b66595c23f61e901764.

Reason: In the end, the dependency on 'kurento-client' is required.
Extra Kurento modules will be generated with this dependency, and that
is OK because their auto-generated Java code does actually depend on the
Kurento Java client classes, so it makes sense that kurento-client is
listed in the dependencies of their pom.xml.

We'll solve the chicken-and-egg problem on the opposite direction: by
removing the extra Kurento modules from
kurento-java/kurento-parent-pom/pom.xml